+++ libcaf_net/caf/internal/accept_handler.cpp
@@ -37,10 +37,6 @@ public:
   error start(net::socket_manager* owner) override {
     auto lg = log::net::trace("");
     owner_ = owner;
-    if (auto err = acceptor_->start(owner); err.valid()) {
-      log::net::debug("failed to start the acceptor: {}", err);
-      return err;
-    }
     self_ref_ = owner->as_disposable();
     if (!monitored_actors_.empty()) {
       monitor_callback_ = make_action([this] { owner_->shutdown(); });
@@ -54,6 +50,10 @@ public:
       }
     }
     on_conn_close_ = make_action([this] { connection_closed(); });
+    if (auto err = acceptor_->start(owner, on_conn_close_); err.valid()) {
+      log::net::debug("failed to start the acceptor: {}", err);
+      return err;
+    }
     owner->register_reading();
     return none;
   }
@@ -81,7 +81,10 @@ public:
                         handle().id);
         owner_->deregister_reading();
       }
-      child->add_cleanup_listener(on_conn_close_);
+      // Note: the acceptor is responsible for ensuring on_conn_close_ is called
+      // when the connection is fully closed. For most protocols, this happens
+      // when the socket manager cleans up. For HTTP, this also requires all
+      // outstanding http::request objects to be destroyed.
       if (auto err = child->start(); err.valid()) {
         on_error(err);
       }