+++ libcaf_net/caf/net/lp/with.cpp
@@ -38,8 +38,9 @@ public:
     // nop
   }
 
-  error start(net::socket_manager* parent) override {
+  error start(net::socket_manager* parent, action on_conn_close) override {
     parent_ = parent;
+    on_conn_close_ = std::move(on_conn_close);
     mcast_ = parent->add_child(std::in_place_type<flow::op::mcast<event_type>>);
     flow::observable<event_type>{mcast_}.subscribe(events_);
     return none;
@@ -79,7 +80,10 @@ public:
                                               std::move(impl));
     transport->max_consecutive_reads(max_consecutive_reads_);
     transport->active_policy().accept();
-    return net::socket_manager::make(parent_->mpx_ptr(), std::move(transport));
+    auto res = net::socket_manager::make(parent_->mpx_ptr(),
+                                         std::move(transport));
+    res->add_cleanup_listener(on_conn_close_);
+    return res;
   }
 
 private:
@@ -96,6 +100,8 @@ private:
   size_t max_message_size_;
 
   size_field_type size_type_;
+
+  action on_conn_close_;
 };
 
 } // namespace